Converted client interact entity packet
